agents/worker-agent/pyproject.toml
[project]
name = "worker-agent"
version = "0.1.0"
description = "Autonomous worker agent that handles full PR lifecycle: implement, review, fix, merge, verify main"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk",
    "pygithub",
    "pydantic>=2.0",
    "rich",
    "typer",
    "python-dotenv",
]

[project.scripts]
worker = "worker_agent.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio",
    "pytest-cov",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM", "TCH", "ASYNC"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v --cov=worker_agent --cov-report=term-missing --cov-fail-under=70"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/worker_agent"]
